Module FormData.Iterator

We need to bind to JavaScript Iterators for FormData functionality. But ideally this binding should be moved into BuckleScript itself.

@see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Iteration_protocols

module Next : sig ... end

type 'a t

val forEach : f:('a -> unit) Js.Fn.arity1 -> 'a t -> unit

forEach ~f iterator runs f on each item returned by the iterator. This is not defined by the platform but a convenience function.

val next : 'a t -> 'a Next.t